graph: separate the absence sentences from the picture machinery
The face held two decisions that change on their own: how the picture is
placed, camera'd and captioned, and what the page SAYS when there is no
picture (the four refusals in zoom's own judgements, the over-ceiling
withholding, the three no-edges arms). The sentences move to
browser/graph/absent.tsx; the face keeps the machinery and reads ONE
derivation for whether there is a picture to draw — the arm attribute
and the branch were two spellings of a single answer, and the drawn list
wraps the face's memo no longer.
