+++ packages/plugins/graph/src/browser/graph/GraphFace.tsx
@@ -21,12 +21,10 @@
  */
 
 import {
-  type Address,
   GRAPH_DRAWN_AT_MOST,
   type Centre,
   HOPS,
   type Hops,
-  printAddress,
   type Shown,
   type Vertex,
 } from "@olai/format"
@@ -41,6 +39,7 @@ import type { Drawn } from "olai-plugin-outlines/page"
 import { TESTID } from "../../testids.ts"
 import { navigationHeld } from "../held.ts"
 import { graph, graphAround } from "../routes.ts"
+import { Edges0, Empty } from "./absent.tsx"
 import { legible, rankedBy } from "./camera.ts"
 import { Canvas, saidOf } from "./Canvas.tsx"
 import { Controls } from "./Controls.tsx"
@@ -128,11 +127,15 @@ export function GraphFace(props: {
     return `every reference in the directory — ${props.page.held} vertices, ${props.page.edges.length} arrows`
   })
 
+  /** The one ruling on whether there is a picture to show at all — the
+   *  arm attribute and the branch below read the SAME answer, never two
+   *  spellings of it. */
+  const drawable = () => refused() === undefined && held() <= GRAPH_DRAWN_AT_MOST
+
   /** Which arm the face drew — the absence assertions in the e2e suite must
    *  never be satisfied by the page that has not yet LANDED: they anchor on
    *  this attribute being present before counting what is not there. */
-  const arm = () =>
-    refused() === undefined && held() <= GRAPH_DRAWN_AT_MOST ? "picture" : "sentence"
+  const arm = () => (drawable() ? "picture" : "sentence")
 
   return (
     <div
@@ -144,7 +147,7 @@ export function GraphFace(props: {
       data-held={held() > 0 ? String(held()) : undefined}
     >
       <Show
-        when={refused() === undefined && held() <= GRAPH_DRAWN_AT_MOST}
+        when={drawable()}
         fallback={
           <Empty
             held={held()}
@@ -213,8 +216,9 @@ function Shape(props: {
   // so this watches the shape rather than the placement.
   createEffect(on(toPlace, looking.fit, { defer: true }))
 
-  /** The drawn dots, as ONE list keyed the way the canvas keys them. */
-  const drawn = createMemo(() => props.drawn())
+  /** The drawn dots, as ONE list keyed the way the canvas keys them —
+   *  already the face's memo, so this is an ALIAS rather than a wrap. */
+  const drawn = props.drawn
 
   /**
    * What stays bright while a reader is pointing at a dot: that dot and
@@ -405,124 +409,6 @@ function CentreHere(props: {
       </Link>
     </Show>
   )
-}
-
-/**
- * The three things the page says instead of drawing an absent centre, and
- * the one it says when the picture is withheld.
- */
-function Empty(props: {
-  readonly held: number
-  readonly reason: Centre | undefined
-}) {
-  return (
-    <div data-testid={TESTID.graphEmpty} data-reason={props.reason?.kind ?? "held"} class="py-6">
-      <Show
-        when={props.reason !== undefined}
-        fallback={
-          <p class="max-w-lg text-ink">
-            The whole reference map of this directory is over the ceiling this
-            page draws under — {props.held} vertices, against {" "}{GRAPH_DRAWN_AT_MOST} — so
-            nothing is drawn: a partial picture of a directory is a sentence
-            with a silent word missing. Narrowing THIS page cannot help —
-            it has nothing on it — but a tighter circle does: open a
-            NEIGHBOURHOOD — a row's ••• has "Reference graph", and the
-            palette has the same page by name.
-          </p>
-        }
-      >
-        <Reason sentence={props.reason} />
-      </Show>
-    </div>
-  )
-}
-
-function Reason(props: { readonly sentence: Centre | undefined }) {
-  switch (props.sentence?.kind) {
-    case "unknown":
-      return props.sentence.address.kind === "node"
-        ? (
-          <p class="max-w-lg text-ink">
-            No node is called <code class="font-mono text-[0.8125rem]">#{props.sentence.address.id}</code> — it was
-            deleted, or the outline holding it is no longer served.
-          </p>
-        )
-        : (
-          <p class="max-w-lg text-ink">
-            No file at <code class="font-mono text-[0.8125rem]">{props.sentence.address.path}</code> — it was
-            moved, renamed, or is no longer one of the served kinds.
-          </p>
-        )
-    case "put-away":
-      return (
-        <p class="max-w-lg text-ink">
-          <NameOf address={props.sentence.address} /> is in the Trash, which is
-          the one page that draws it.{" "}
-          <Link route={{ kind: "trash" }} class="underline">Open the Trash</Link>.
-        </p>
-      )
-    case "dangling":
-      return (
-        <p class="max-w-lg text-ink">
-          <code class="font-mono text-[0.8125rem]">#{props.sentence.id}</code> is a mirror, and the chain from it
-          ends at <code class="font-mono text-[0.8125rem]">{props.sentence.missing}</code>, which no node
-          declares — so there is no vertex the graph could centre on.
-        </p>
-      )
-    case "cycle":
-      return (
-        <p class="max-w-lg text-ink">
-          <code class="font-mono text-[0.8125rem]">#{props.sentence.id}</code> is a mirror whose chain closes on
-          itself at <code class="font-mono text-[0.8125rem]">{props.sentence.through}</code>, so it stands for
-          no vertex at all.
-        </p>
-      )
-    default:
-      return null
-  }
-}
-
-/** The address in the sentence, said the way the URL would say it — ONE
- *  spelling, so a sentence and a link never say two names for one thing. */
-function NameOf(props: { readonly address: Address }) {
-  return <code class="font-mono text-[0.8125rem]">{printAddress(props.address)}</code>
-}
-
-/** The sentences drawn where a picture would be — one per way "no edges"
- *  happens, each in the words of the arm the reader is in:
- *  - nothing ever referred anywhere (the open map, unfiltered);
- *  - the FILTER took everything — a sentence about the filter, never a
- *    claim about the directory (a head-counted picture stays a picture);
- *  - nothing refers to the centre, and it refers to nothing — the CENTRE
- *    NAMED, because a sentence reading "this one" is a sentence from the
- *    perspective nobody holds.
- */
-function Edges0(props: {
-  readonly page: Extract<Shown, { readonly kind: "graph" }>
-  readonly drawn: Pick<Extract<Drawn, { readonly kind: "graph" }>, "vertices">
-  | undefined
-}) {
-  return (
-    <p class="max-w-lg text-ink" data-testid={TESTID.graphEmpty} data-reason="no-edges">
-      {props.page.around === null
-        ? (props.drawn?.vertices.length ?? 0) === 0 && props.page.vertices.length > 0
-          ? <>The filter took every dot on this map — say less, or clear it, to see the picture.</>
-          : <>Nothing in this directory refers to anything yet — no node points at another with <code class="font-mono text-[0.8125rem]">see</code>, no outline is another record's <code class="font-mono text-[0.8125rem]">doc</code>, and no note names a vertex at all.</>
-        : props.page.hops === 1
-        ? <>Nothing refers to <NameTitle page={props.page} />, and it refers to nothing — no <code class="font-mono text-[0.8125rem]">see</code> written out and none pointed back.</>
-        : <>Nothing refers to <NameTitle page={props.page} />, and it refers to nothing — not even two hops out.</>}
-    </p>
-  )
-}
-
-/** The centre's own name in a sentence — read off the most reliable name the
- *  page holds: the resolved centre's vertex first, the bare address print
- *  when there was none. */
-function NameTitle(props: { readonly page: Extract<Shown, { readonly kind: "graph" }> }) {
-  const around = () => props.page.around
-  return around()?.kind === "vertex"
-    ? <>{(around() as { kind: "vertex"; vertex: Vertex }).vertex.title}</>
-    : null
 }
 
 /**
